raster-to-vector/server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
numpy>=1.26.0
opencv-python-headless>=4.9.0
scikit-image>=0.22.0
scipy>=1.12.0
Pillow>=10.2.0
cairosvg>=2.7.0
svgwrite>=1.4.3
svgpathtools>=1.6.1
pydantic>=2.5.0
